capstone-submission/CHN/Karthik K/CapstoneProject-KarthikKamalakannan/requirements_server.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
langchain-core>=0.3
langchain-groq>=0.2
langgraph>=0.2
pymupdf>=1.24
python-dotenv>=1.0
arize-phoenix
arize-phoenix-otel
opentelemetry-sdk
